requirements.txt
# MCP Server
mcp>=0.9.0

# API Client
openai>=1.0.0

# PDF Processing
pymupdf>=1.23.0

# Image Processing
Pillow>=10.0.0

# Environment Variables
python-dotenv

# YAML Configuration
PyYAML

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0

# Web API
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
